pyproject.toml
[tool.poetry]
name = "ai-billing-backend"
version = "0.1.0"
description = "AI-Powered SaaS Usage & Billing Backend"
authors = ["Antigravity <antigravity@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
redis = "^5.0.1"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.1.0"
transformers = "^4.37.0"
torch = "^2.1.0" # Large dependency!

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
